requirements.txt
django-resized
asgiref==3.2.10
certifi==2020.6.20
cycler==0.10.0
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
django-resized
future==0.18.2
gunicorn==20.0.4
imageio==2.9.0
kiwisolver==1.2.0
matplotlib==3.3.2
networkx==2.5
numpy==1.19.2
Pillow==7.2.0
psycopg2==2.8.6
Pympler==0.9
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.2
six==1.15.0
sqlparse==0.4.1
djangorestframework
tifffile
whitenoise==5.2.0
wincertstore==0.2
opencv-python-headless
django-storages
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.6.0+cpu
torchvision==0.7.0+cpu


